--- includes/userClass.php
@@ -1,11 +1,0 @@
-<?php
-class User {
-    public $email;
-    public $password;
-    function __construct($email, $password) {
-        $this->email = $email;
-        $this->password = $password;
-    }
-} 
-
-?>
